feature: ADD accout/model object
